Add drag drawing unit test

+++ utils/common/include/proto.h
@@ -30,7 +30,7 @@ static const int32_t INVALID_PID = 7;
 static const int32_t SESSION_NOT_FOUND = 8;
 static const int32_t EPOLL_MODIFY_FAIL = 9;
 static const int32_t ADD_SESSION_FAIL = 11;
-static const int32_t MAX_SESSON_ALARM = 100;
+static const size_t MAX_SESSON_ALARM = 100;
 static const int32_t MAX_RECV_LIMIT = 13;
 static const int32_t SERVICE_NOT_RUNNING = 14;
 #define CONNECT_MODULE_TYPE_FI_CLIENT 0
